backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "celery>=5.6.2",
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
    "uvicorn-worker>=0.4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true